moved Brew.js into Brews component folder
